Initial setup.py config
Adding an initial setup.py and setup.cfg configuration, to be able to
automate installation, packaging and distribution.

A good packaging and distribution solution is required if we want to
publish to PyPI, and improve the installation process for end users.
It will also simplify the inclusion of other libraries as dependencies.